+++ ricaun.Revit.UI/RibbonItemDataExtension.cs
@@ -3,6 +3,9 @@ using System.Windows.Media;
 
 namespace ricaun.Revit.UI
 {
+    /// <summary>
+    /// RibbonItemDataExtension
+    /// </summary>
     public static class RibbonItemDataExtension
     {
 
@@ -119,9 +122,13 @@ namespace ricaun.Revit.UI
             {
                 if (ribbonItem is ButtonData ribbonButton)
                 {
-                    ribbonButton.LargeImage = largeImage;
+                    ribbonButton.LargeImage = largeImage.GetBitmapFrame(32);
                     if (ribbonButton.Image == null)
-                        ribbonButton.Image = largeImage.Scale(0.5);
+                    {
+                        ribbonButton.Image = largeImage.GetBitmapFrame(16);
+                        if (ribbonButton.Image.Width != 16)
+                            ribbonButton.Image = ribbonButton.Image.Scale(16 / ribbonButton.Image.Width);
+                    }
                 }
             }